pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["streetscapes"]

[project]
name = "streetscapes"
version = "2024"
dependencies = [
    "black",
    "cloup",
    "dask",
    "ephem",
    "fiona",
    "geodatasets",
    "geopandas",
    "h3",
    "ipykernel",
    "mapillary @ git+https://github.com/Urban-M4/mapillary-python-sdk.git",
    "matplotlib",
    "numpy",
    "osmnx",
    "pandas",
    "pyarrow",
    "pyproj",
    "python-decouple",
    "rasterio",
    "seaborn",
    "timezonefinder",
    "tqdm",
]
authors = []
description = "Code for extracting the emissivity and albedo of buildings in cities based on images downloaded from Mapillary and KartaView."
readme = "README.md"
license = { file = "LICENCE.md" }


[project.scripts]
streetscapes = "streetscapes.main:main"
